backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.6

# Database
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
alembic>=1.13.2
pgvector>=0.3.2

# Pydantic (with email validation)
pydantic[email]>=2.9.2
pydantic-settings>=2.5.2

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Redis & HTTP Client
redis>=5.0.8
httpx>=0.27.2

# File handling & utilities
python-multipart>=0.0.9
python-dotenv>=1.0.1

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
python-magic>=0.4.27

# Google AI
google-genai>=1.0.0

# Logging (optional - you can remove if not using structured logs)
structlog>=23.2.0
